feat(language-server): support `--tsdk`` command line option
