+++ extensions/vscode/rolldown.config.ts
@@ -44,7 +44,7 @@ const config: RolldownOptions = {
 					fs.mkdirSync(resolve('./dist'), { recursive: true });
 					fs.writeFileSync(
 						resolve('./dist/language-server.js'),
-						`module.exports = require('../node_modules/@vue/language-server/index.js');`,
+						`module.exports = require('../node_modules/@vue/language-server/bin/vue-language-server.js');`,
 					);
 					fs.writeFileSync(
 						resolve('./dist/typescript-plugin.js'),
@@ -74,7 +74,7 @@ const config: RolldownOptions = {
 					id: /^typescript$/,
 				},
 				handler() {
-					return { id: './typescript.js', external: true };
+					return { id: 'typescript', external: true };
 				},
 			},
 		},
@@ -84,7 +84,7 @@ const config: RolldownOptions = {
 if (!isDev) {
 	config.input = {
 		...config.input as Record<string, string>,
-		'language-server': './node_modules/@vue/language-server/index.js',
+		'language-server': './node_modules/@vue/language-server/bin/vue-language-server.js',
 		'typescript-plugin': './node_modules/@vue/typescript-plugin/index.js',
 	};
 }
