+++ packages/language-server/bin/vue-language-server.js
@@ -5,4 +5,15 @@ if (process.argv.includes('--version')) {
 	return;
 }
 
-require('../index.js');
+let ts;
+for (const arg of process.argv) {
+	if (arg.startsWith('--tsdk=')) {
+		const tsdk = arg.substring('--tsdk='.length);
+		const tsPath = require.resolve('./typescript.js', { paths: [tsdk] });
+		ts = require(tsPath);
+		break;
+	}
+}
+ts ??= require('typescript');
+
+require('../lib/server').startServer(ts);
